+++ README.md
@@ -1268,6 +1268,9 @@ cmd输入命令即可：`doctoc /path/to/file`
 - 定时任务 报时任务补充分钟的播报
 - 修复定时任务同时存在多个变量时，无法正常运行的bug
 
+### 2023-08-04
+- 更换快手监听方案，新方案是playwright的形式，建议使用小号，每次用完需要删除cookie下的json文件，因为第二次就异常了，没法正常捕获。暂时只支持弹幕回复。
+
 </details>
